get rid of generic connection update window function
