+++ source/h2_connection.c
@@ -61,7 +61,7 @@ static struct aws_http_stream *s_connection_make_request(
     const struct aws_http_make_request_options *options);
 static void s_connection_close(struct aws_http_connection *connection_base);
 static bool s_connection_is_open(const struct aws_http_connection *connection_base);
-static void s_connection_update_window(struct aws_http_connection *connection_base, size_t increment_size);
+static int s_connection_update_window(struct aws_http_connection *connection_base, size_t increment_size);
 static int s_connection_change_settings(
     struct aws_http_connection *connection_base,
     const struct aws_http2_setting *settings_array,
@@ -163,7 +163,7 @@ static struct aws_http_connection_vtable s_h2_connection_vtable = {
     .stream_send_response = NULL,
     .close = s_connection_close,
     .is_open = s_connection_is_open,
-    .update_window = s_connection_update_window,
+    .h2_update_window = s_connection_update_window,
     .change_settings = s_connection_change_settings,
     .send_ping = s_connection_send_ping,
     .send_goaway = s_connection_send_goaway,
@@ -2097,16 +2097,17 @@ static bool s_connection_is_open(const struct aws_http_connection *connection_ba
     return is_open;
 }
 
-static void s_connection_update_window(struct aws_http_connection *connection_base, size_t increment_size) {
+static int s_connection_update_window(struct aws_http_connection *connection_base, size_t increment_size) {
     struct aws_h2_connection *connection = AWS_CONTAINER_OF(connection_base, struct aws_h2_connection, base);
     if (!increment_size) {
-        return;
+        CONNECTION_LOG(TRACE, connection, "Ignoring window update of size 0.");
+        return AWS_OP_SUCCESS;
     }
     if (!connection_base->manual_window_management) {
         /* auto-mode, manual update window is not supported */
         CONNECTION_LOG(
             WARN, connection, "Manual window management is off, update window operations are not supported.");
-        return;
+        return AWS_OP_SUCCESS;
     }
     /* Type cast the increment size here, if overflow happens, we will detect it later, and the frame will be destroyed
      */
@@ -2118,7 +2119,7 @@ static void s_connection_update_window(struct aws_http_connection *connection_ba
             connection,
             "Failed to create WINDOW_UPDATE frame on connection, error %s",
             aws_error_name(aws_last_error()));
-        return;
+        return AWS_OP_ERR;
     }
 
     int err = 0;
@@ -2149,9 +2150,8 @@ static void s_connection_update_window(struct aws_http_connection *connection_ba
 
     if (!connection_open) {
         CONNECTION_LOG(ERROR, connection, "Failed to update connection window, connection is closed or closing.");
-        aws_raise_error(AWS_ERROR_INVALID_STATE);
         aws_h2_frame_destroy(connection_window_update_frame);
-        return;
+        return aws_raise_error(AWS_ERROR_INVALID_STATE);
     }
 
     if (err) {
@@ -2163,10 +2163,11 @@ static void s_connection_update_window(struct aws_http_connection *connection_ba
             "The increment size is too big for HTTP/2 protocol, max flow-control "
             "window size is 2147483647. We got %zu, which will cause the flow-control window to exceed the maximum",
             increment_size);
-        aws_raise_error(AWS_ERROR_INVALID_ARGUMENT);
         aws_h2_frame_destroy(connection_window_update_frame);
-        return;
+        return aws_raise_error(AWS_ERROR_INVALID_ARGUMENT);
     }
+
+    return AWS_OP_SUCCESS;
 }
 
 static int s_connection_change_settings(
